Cleanup racetest targets
Right now we have targets for race test and unit test. In CI, we only
run race tests, so its not clear there is much benefit to having the
non-race test variants, especially since they have caused multiple bugs
in the past: https://github.com/istio/istio/issues/19559 and
https://github.com/istio/istio/issues/20110 as examples.

This makes test and racetest both use -race. In a followup PR, after CI
is changed, we will just remove the racetest target and only use test
for simplicity.

Additionally, rather than expecting RACE_TEST to be set, use native go
support for excluding tests with -race passed. Also made a change to
mixer tests to properly default REPO_ROOT. All of these changes all a
user to just run `go test ./(anything but e2e tests)` and things just
work, without relying on Makefiles.
