+++ mixer/adapter/kubernetesenv/kubernetesenv_test.go
@@ -527,47 +527,6 @@ func mockCreateInterfaceFromClusterConfig(_ *clientcmdapi.Config) (kubernetes.In
 	return fake.NewSimpleClientset(), nil
 }
 
-func Test_KubeSecretController(t *testing.T) {
-	if len(os.Getenv("RACE_TEST")) > 0 {
-		t.Skip("https://github.com/istio/istio/issues/15610")
-	}
-
-	secretcontroller.LoadKubeConfig = mockLoadKubeConfig
-	secretcontroller.ValidateClientConfig = mockValidateClientConfig
-	secretcontroller.CreateInterfaceFromClusterConfig = mockCreateInterfaceFromClusterConfig
-
-	clientset := fake.NewSimpleClientset()
-	b := newBuilder(func(string, adapter.Env) (kubernetes.Interface, error) {
-		return clientset, nil
-	})
-
-	// Call kube Build function which will start the secret controller.
-	// Sleep to allow secret process to start.
-	_, err := b.Build(context.Background(), test.NewEnv(t))
-	if err != nil {
-		t.Fatalf("error building adapter: %v", err)
-	}
-	time.Sleep(10 * time.Millisecond)
-
-	// Create the multicluster secret.
-	err = createMultiClusterSecret(clientset)
-	if err != nil {
-		t.Fatalf("Unexpected error on secret create: %v", err)
-	}
-
-	// Test - Verify that the remote controller has been added.
-	verifyControllers(t, b, 2, "create remote controller")
-
-	// Delete the mulicluster secret.
-	err = deleteMultiClusterSecret(clientset)
-	if err != nil {
-		t.Fatalf("Unexpected error on secret delete: %v", err)
-	}
-
-	// Test - Verify that the remote controller has been removed.
-	verifyControllers(t, b, 1, "delete remote controller")
-}
-
 // Kubernetes Runtime Object for Tests
 
 var trueVar = true
