+++ mixer/test/spybackend/nosession_integration_test.go
@@ -1,3 +1,5 @@
+// +build !race
+
 // Copyright 2018 Istio Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -17,7 +19,6 @@ package spybackend
 import (
 	"fmt"
 	"io/ioutil"
-	"os"
 	"strings"
 	"testing"
 	"time"
@@ -254,13 +255,6 @@ spec:
 )
 
 func TestNoSessionBackend(t *testing.T) {
-
-	// We skip this test if it is being run as part of the racetest because
-	// it is intensive on memory and will cause the test to fail with no error
-	if len(os.Getenv("RACE_TEST")) > 0 {
-		t.Skip()
-	}
-
 	testdata := []struct {
 		name   string
 		calls  []adapter_integration.Call
