--- tests/util/kubeconfig
@@ -1,12 +1,0 @@
-apiVersion: v1
-clusters:
-- cluster:
-    server: http://localhost:8080
-  name: istio
-contexts:
-- context:
-    cluster: istio
-    user: ""
-  name: istio
-current-context: istio
-kind: Config
